requirements.txt
# Make sure to install those as additional_dependencies in the
# pre-commit hooks for pylint & mypy
beautifulsoup4
thefuzz==0.19.0
python-Levenshtein
python-telegram-bot==13.7
Sphinx
requests
github3.py==2.0.0
